Redo volume management to be present
Previously to this change, jobs would stay alive to provide their
volumes to be used by other jobs. When that happened, the requesting job
would ask atc for the volume, and atc would ask the ownering job to
stream the volume back, where atc would stream the data directly back to
the requesting job.

There are a number of flaws with this approach:
* Any volume that is used more that once is requested from the owner job
  multiple times, redundently sending the data over the wire to atc.
* It requires the owning job to stay alive for the rest of the build.
  This is the biggest flaw because it means that the resources used by
  these owning jobs is not released until the whole build finishes.
  This easily causes resource starvation and builds stall as the nomad
  cluster can not schedule the requesting jobs and everything just
  grinds to a halt.

This new approach has jobs, once their command is finished, stream their
volumes back to atc immediately. atc then writes those volumes (which is
a snappy compressed tar file) to a work directory under the build number
and using the volume handle. Once a job has sent all it's volumes, it
exits which frees up it's resources for new jobs to run.

Any jobs that need volumes generated by other jobs calls the same
RequestVolume call at the start like it did before, but now atc serves
the data out of it's work directory rather than requesting it from the
owning job.

When the build finishs, atc wipes the build specific directory
containing all the volumes.